Made some type annotation changes

+++ src/main/java/entities/MealPlan.java
@@ -5,12 +5,13 @@ import java.util.List;
 public class MealPlan {
 
     private final List<Recipe> recipes;
-    private final int targetCalories;
-    private final int targetProtein;
-    private final int targetCarbs;
-    private final int targetFats;
+    private final double targetCalories;
+    private final double targetProtein;
+    private final double targetCarbs;
+    private final double targetFats;
 
-    public MealPlan(List<Recipe> recipes, int targetCalories, int targetProtein, int targetCarbs, int targetFats) {
+    public MealPlan(List<Recipe> recipes, double targetCalories, double targetProtein, double targetCarbs,
+                    double targetFats) {
         this.recipes = recipes;
         this.targetCalories = targetCalories;
         this.targetProtein = targetProtein;
@@ -18,17 +19,17 @@ public class MealPlan {
         this.targetFats = targetFats;
     }
 
-    public int getTargetCalories() {
+    public double getTargetCalories() {
         return this.targetCalories;
     }
 
-    public int getTargetProtein() {
+    public double getTargetProtein() {
         return this.targetProtein;
     }
 
-    public int getTargetCarbs() { return this.targetCarbs; }
+    public double getTargetCarbs() { return this.targetCarbs; }
 
-    public int getTargetFats() { return this.targetFats; }
+    public double getTargetFats() { return this.targetFats; }
 
     public List<Recipe> getRecipes() { return this.recipes; }
 }
